+++ tests/testthat/test_plotSpectraPTM.R
@@ -2,6 +2,8 @@
 #' # that could be called by `vdiffr::expect_doppelganger()`
 #' Run devtools::test_active_file(file = "tests/testthat/test_plotSpectraPTM.R")
 
+library("Spectra")
+
 sp <- DataFrame(
     msLevel = 2L,
     rtime = 2345,
@@ -10,12 +12,16 @@ sp <- DataFrame(
     scanIndex = 1L,
     charge = 2L
 )
-sp$mz <- list(c(223.1583, 251.15432, 308.168017, 455.24801, 604.30949,
-                641.30842, 667.2244, 778.30164, 813.34935, 923.350391,
-                995.45281, 1017.43394, 1065.46197, 1112.5069, 1130.5874))
-sp$intensity <- list(c(83000, 65000, 190000, 379000, 281000, 112000, 39000,
-                       139000, 1015000, 63000, 58000, 1960000, 240000,
-                       1338000, 40700))
+sp$mz <- list(c(
+    223.1583, 251.15432, 308.168017, 455.24801, 604.30949,
+    641.30842, 667.2244, 778.30164, 813.34935, 923.350391,
+    995.45281, 1017.43394, 1065.46197, 1112.5069, 1130.5874
+))
+sp$intensity <- list(c(
+    83000, 65000, 190000, 379000, 281000, 112000, 39000,
+    139000, 1015000, 63000, 58000, 1960000, 240000,
+    1338000, 40700
+))
 spectra <- Spectra(sp)
 
 test_that("plotSpectraPTM works with deltaMz = TRUE", {
@@ -25,7 +31,8 @@ test_that("plotSpectraPTM works with deltaMz = TRUE", {
             plotSpectraPTM(
                 spectra,
                 type = c("a", "b", "c", "x", "y", "z"),
-                deltaMz = TRUE
+                deltaMz = TRUE,
+                z = 1
             )
         }
     )
@@ -38,7 +45,8 @@ test_that("plotSpectraPTM works with deltaMz = FALSE", {
             plotSpectraPTM(
                 spectra,
                 type = c("a", "b", "c", "x", "y", "z"),
-                deltaMz = FALSE
+                deltaMz = FALSE,
+                z = 1
             )
         }
     )
@@ -67,7 +75,8 @@ test_that("plotSpectraPTM works with different col", {
                 spectra,
                 col = c(y = "red", b = "blue", acxy = "orange", other = "violet"),
                 type = c("a", "b", "c", "x", "y", "z"),
-                deltaMz = FALSE
+                deltaMz = FALSE,
+                z = 1
             )
         }
     )
@@ -80,7 +89,8 @@ test_that("plotSpectraPTM works with USI = FALSE", {
             plotSpectraPTM(
                 spectra,
                 type = c("a", "b", "c", "x", "y", "z"),
-                USI = FALSE
+                USI = FALSE,
+                z = 1
             )
         }
     )